moved dialog to shared so that repl order history can use it as well
