+++ projects/storefrontapp/src/environments/environment.ts
@@ -8,12 +8,12 @@ import { Environment } from './models/environment.model';
 export const environment: Environment = {
   production: false,
   occBaseUrl:
-    'https://spartacus-dev0.eastus.cloudapp.azure.com:9002' ||
+    'https://localhost:9002' ||
     // 'https://spartacus-dev3.eastus.cloudapp.azure.com:9002' ||
     // 'https://api.c39j2-walkersde1-d4-public.model-t.cc.commerce.ondemand.com' ||
     build.process.env.SPARTACUS_BASE_URL,
   occApiPrefix: '/occ/v2/',
-  b2b: false,
+  b2b: true,
   cds: false,
   cdc: false,
 };
