+++ projects/storefrontlib/src/cms-components/myaccount/order/replenishment-order-details/default-replenishment-order-cancellation-layout.config.ts
@@ -1,6 +1,6 @@
 import { LayoutConfig } from '../../../../layout/config/layout-config';
 import { DIALOG_TYPE } from '../../../../layout/launch-dialog/index';
-import { ReplenishmentOrderCancellationDialogComponent } from './replenishment-order-cancellation/index';
+import { ReplenishmentOrderCancellationDialogComponent } from '../../../../shared/components/replenishment-order-cancellation-dialog/replenishment-order-cancellation-dialog.component';
 
 export const defaultReplenishmentOrderCancellationLayoutConfig: LayoutConfig = {
   launch: {
