+++ projects/storefrontlib/src/cms-components/myaccount/order/replenishment-order-details/index.ts
@@ -1,4 +1,4 @@
 export * from './default-replenishment-order-cancellation-layout.config';
-export * from './replenishment-order-cancellation/index';
+export * from './replenishment-order-cancellation/replenishment-order-cancellation.component';
 export * from './replenishment-order-details.module';
 export * from './replenishment-order-details.service';
