--- projects/storefrontlib/src/cms-components/myaccount/order/replenishment-order-details/replenishment-order-cancellation/index.ts
@@ -1,2 +1,0 @@
-export * from './replenishment-order-cancellation-dialog/replenishment-order-cancellation-dialog.component';
-export * from './replenishment-order-cancellation.component';
