+++ projects/storefrontlib/src/cms-structure/routing/default-routing-config.ts
@@ -73,6 +73,9 @@ export const defaultStorefrontRoutesConfig: RoutesConfig = {
     paths: ['my-account/coupon/claim/:couponCode'],
     paramsMapping: { couponCode: 'code' },
   },
+  replenishmentOrders: {
+    paths: ['my-account/my-replenishment'],
+  },
   replenishmentDetails: {
     paths: ['my-account/my-replenishment/:replenishmentOrderCode'],
     paramsMapping: { replenishmentOrderCode: 'code' },
