+++ projects/storefrontlib/src/shared/components/index.ts
@@ -8,6 +8,7 @@ export * from './list-navigation/index';
 export * from './media/index';
 export * from './modal/index';
 export * from './order-overview/index';
+export * from './replenishment-order-cancellation-dialog/index';
 export * from './spinner/index';
 export * from './split-view/index';
 export * from './star-rating/index';
