+++ projects/storefrontlib/src/shared/components/replenishment-order-cancellation-dialog/index.ts
@@ -1,0 +1,2 @@
+export * from './replenishment-order-cancellation-dialog.component';
+export * from './replenishment-order-cancellation-dialog.module';
