+++ projects/storefrontlib/src/shared/components/replenishment-order-cancellation-dialog/replenishment-order-cancellation-dialog.module.ts
@@ -1,0 +1,13 @@
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+import { I18nModule } from '@spartacus/core';
+import { IconModule } from '../../../cms-components/misc/icon/icon.module';
+import { ReplenishmentOrderCancellationDialogComponent } from './replenishment-order-cancellation-dialog.component';
+
+@NgModule({
+  imports: [CommonModule, I18nModule, IconModule],
+  declarations: [ReplenishmentOrderCancellationDialogComponent],
+  entryComponents: [ReplenishmentOrderCancellationDialogComponent],
+  exports: [ReplenishmentOrderCancellationDialogComponent],
+})
+export class ReplenishmentOrderCancellationDialogModule {}
